requirements.txt
torch
torchvision
pynndescent
llvmlite>=0.34.0
sklearn
umap-learn
torchsummary
